Remove createPairs function and direct assignment of data arrays
Removed the createPairs function from the variables.js file and directly assigned the corresponding data arrays to the variables source1magnitude and source1phase. This change was done to simplify the code and make the way the data is structured and stored more evident and easier to understand. Furthermore, the removed function was only used in this specific case so removing it does not affect other parts of the code.