Reduced frequency data used and added crosshair in plot
The frequency arrays for 'SUM' and 'TARGET' in `variables.js` have been significantly reduced for efficiency and readability. This simplifies the data structure and potentially enhances the performance of functions consuming this data.

Moreover, additional features, like a crosshair, have been added to the plot in `main.js` to improve the user experience by facilitating precise reading of data points on plot.