Remove hardcoded frequency and magnitude data in variables.js
The hardcoded "transferFunctions" data for frequencies and magnitudes in variables.js were stripped out. This is part of the larger move to load this and similar data from more maintainable external data files or services, instead of having it all hardcoded. The change simplifies the code making it easier to read, maintain, and prevents potential errors due to manual data manipulation.