Refactor condition checks and value assignment
The previously declared `sources` variable was being checked for null. This was an unnecessary check as `sources` would never be null - it is either undefined or has a value. Hence it was removed.

Also, there was a minor deviation in code styling: there was no space after the colon in the object literals right before the `corridor60degStart`, `corridor60degEnd`, `xovrStart`, `xovrEnd` properties. This was fixed for better readability and consistency with common JavaScript code style.